backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
python-dateutil==2.8.2
openai==0.28.0
python-dotenv==1.0.0